requirements.txt
azure-functions
azure-storage-blob
google-cloud-storage
google-auth
msal
requests
